Stable Commit:
    UI Update:
    - add card fragment implemented with basic feature
    - navigation refinements
    - color palette improved
    - styles reorganized
    - animations added between fragment transitions
    - new menu added for different fragments (filter, edit/add credentials) etc.

    Core Update:
    - validation class updated
    - card form validation added
    - pin data type changed from integer to short

    Test Update:
    - no changes

+++ app/src/main/java/com/th3pl4gu3/locky/core/Card.kt
@@ -9,7 +9,7 @@ data class Card(
     val id: String = "",
     var name: String = "",
     var number: Long = 0,
-    var pin: Int = 0,
+    var pin: Short = 0,
     var bank: String = "",
     var cardHolderName: String = "",
     var issuedDate: Calendar = Calendar.getInstance(),
